Fix Preview component overflowing outer div
Reason for this is because I applied a h-96 height to the div wrapping
Editor, PreviewBtn and Preview. So this caused it to stay at the fixed
height while the Preview component grew out.

Also made the styles in the Editor and Preview consistent. In Editor
textarea had h-96 while in Preview the outer wrapping div had h-96
instead of the inner preview displaying div (div equivalent of textarea
in Editor). Made both consistent.