Replace BrowserRouter with HashRouter
Did this to prevent the problem of not being able to copy or manually
enter the url in a new tab and load the site. This was caused due to the
fact the /pages/:id if not a file present in the server and when copied
or manually entered on the site, a request goes to the server for the
file /pages/:id. Instead we need to load the baseurl first then let
react route to /pages/:id.